% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CastorCSV.R
\docType{package}
\name{CastorCSV}
\alias{CastorCSV}
\title{CastorCSV: A package for reading in csv exports from Castoredc.}
\description{
The CastorCSV package provides functions for reading and cleaning
Castoredc csv exports.
}
