% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_variables_option_groups.R
\name{map_variables_option_groups}
\alias{map_variables_option_groups}
\title{map_variables_option_groups}
\usage{
map_variables_option_groups(
  data,
  variable_list,
  field_options,
  checkbox_options = c("No", "Yes")
)
}
\arguments{
\item{data}{The datasets, in the data_list from the read_zipped_folder output,
which you want to map the values and option groups.}

\item{variable_list}{The variable list. Depending on the data param, choose
report_variablelist for report data, study_variablelist for study data or
survey_variablelist for survey data.
data.}

\item{field_options}{The field options data in the data_list output from
read_zipped_folder.}

\item{checkbox_options}{Choose the language options for the checkboxes,
for example: c("No", "Yes") or c("Nee", "Ja"). In Castoredc the negative
option is 0 and the positive 1. Please mention the negative option first
and then the positive.}
}
\description{
\code{map_variables_option_groups}
}
\examples{
\dontrun{
lapply(data_list[c("study_data")],
       map_variables_option_groups,
       data_list$study_variablelist,
       data_list$field_options)
}
}
