% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zipped_folder.R
\name{read_zipped_folder}
\alias{read_zipped_folder}
\title{read_zipped_folder}
\usage{
read_zipped_folder(export_path)
}
\arguments{
\item{export_path_path}{Path to the Castor export_path zip.}
}
\description{
\code{read_zipped_folder}
}
\examples{
\dontrun{
data_list <- read_zipped_folder(export_path_path)
}
}
